Admin controller and Layout added

+++ BeerShop/Controllers/ItemsController.cs
@@ -15,7 +15,7 @@ namespace BeerShop.Controllers
     public class ItemsController : Controller
     {
         private BeerShopContext db = new BeerShopContext();
-
+        private bool isWOrker = false;
         //
         // GET: /Items/
 
@@ -164,15 +164,20 @@ namespace BeerShop.Controllers
 
         public ActionResult Create()
         {
-            Dictionary<string, SelectList> categoriesDictionary = new Dictionary<string, SelectList>();
-            foreach (var categoryType in db.Categories)
+            if (User is Worker || isWOrker == true) //TODO change when worker will come
             {
-                SelectList SelectCategoryList = new SelectList(categoryType.categories, "CategoryItemID", "name");
-                categoriesDictionary.Add(categoryType.name, SelectCategoryList);
-            }
+                Dictionary<string, SelectList> categoriesDictionary = new Dictionary<string, SelectList>();
+                foreach (var categoryType in db.Categories)
+                {
+                    SelectList SelectCategoryList = new SelectList(categoryType.categories, "CategoryItemID", "name");
+                    categoriesDictionary.Add(categoryType.name, SelectCategoryList);
+                }
 
-            ViewBag.typesList = categoriesDictionary;
-            return View();
+                ViewBag.typesList = categoriesDictionary;
+                return View();
+            }
+            else
+                return RedirectToAction("Index", "Home", new { message = "only worker can do this" });
         }
 
         //
@@ -181,27 +186,29 @@ namespace BeerShop.Controllers
         [HttpPost]
         public ActionResult Create(ItemCategoryHelper itemHelper)
         {
-            
-
-
-            if (ModelState.IsValid)
+            if (User is Worker || isWOrker == true)
             {
-                itemHelper.item.imageURL = GetItemPicture(itemHelper.item.name);
-                itemHelper.item.isStillOnSale = true;
-                db.Items.Add(itemHelper.item);
-                foreach (var selectedCategory in itemHelper.categoryTypeCategoryDictionary)
+                if (ModelState.IsValid)
                 {
+                    itemHelper.item.imageURL = GetItemPicture(itemHelper.item.name);
+                    itemHelper.item.isStillOnSale = true;
+                    db.Items.Add(itemHelper.item);
+                    foreach (var selectedCategory in itemHelper.categoryTypeCategoryDictionary)
+                    {
 
-                    if (selectedCategory.Value.Equals("-1"))
-                        break;
-                    int selectedCategoryID = int.Parse(selectedCategory.Value);
-                    db.CategoryItems.FirstOrDefault(c => c.CategoryItemID == selectedCategoryID).items.Add(itemHelper.item);
+                        if (selectedCategory.Value.Equals("-1"))
+                            break;
+                        int selectedCategoryID = int.Parse(selectedCategory.Value);
+                        db.CategoryItems.FirstOrDefault(c => c.CategoryItemID == selectedCategoryID).items.Add(itemHelper.item);
+                    }
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
                 }
-                db.SaveChanges();
-                return RedirectToAction("Index");
-            }
 
-            return View(itemHelper.item);
+                return View(itemHelper.item);
+            }
+            else
+                return RedirectToAction("Index", "Home", new { message = "only worker can do this" });
         }
 
         private static string GetItemPicture(String name)
@@ -235,6 +242,8 @@ namespace BeerShop.Controllers
 
         public ActionResult Edit(int id = 0)
         {
+            if(User is Worker || isWOrker == true)
+            {
             Item item = db.Items.Find(id);
             if (item == null)
             {
@@ -266,6 +275,9 @@ namespace BeerShop.Controllers
 
             return View(itemHelper);
         }
+            else
+                return RedirectToAction("Index", "Home", new { message = "only worker can do this" });
+        }
 
         //
         // POST: /Items/Edit/5
@@ -273,27 +285,32 @@ namespace BeerShop.Controllers
         [HttpPost]
         public ActionResult Edit(ItemCategoryHelper itemHelper)
         {
-            if (ModelState.IsValid)
+            if (User is Worker || isWOrker)
             {
-                db.Entry(itemHelper.item).State = EntityState.Modified;
-
-                var itemTMP = db.Items.Find(itemHelper.item.ItemID);
-                db.Entry(itemTMP).Collection(i => i.categories).Load();
-
-                itemTMP.categories.ToList().ForEach(cat => itemTMP.categories.Remove(cat));
-
-                foreach (var selectedCategory in itemHelper.categoryTypeCategoryDictionary)
+                if (ModelState.IsValid)
                 {
-                    if (selectedCategory.Value.Equals("-1"))
-                        continue;
-                    int selectedCategoryID = int.Parse(selectedCategory.Value);
-                    db.CategoryItems.FirstOrDefault(c => c.CategoryItemID == selectedCategoryID).items.Add(itemHelper.item);
-                }
-                db.SaveChanges();
-                return RedirectToAction("Details", new { id = itemHelper.item.ItemID });
-            }
+                    db.Entry(itemHelper.item).State = EntityState.Modified;
 
-            return View(itemHelper.item);
+                    var itemTMP = db.Items.Find(itemHelper.item.ItemID);
+                    db.Entry(itemTMP).Collection(i => i.categories).Load();
+
+                    itemTMP.categories.ToList().ForEach(cat => itemTMP.categories.Remove(cat));
+
+                    foreach (var selectedCategory in itemHelper.categoryTypeCategoryDictionary)
+                    {
+                        if (selectedCategory.Value.Equals("-1"))
+                            continue;
+                        int selectedCategoryID = int.Parse(selectedCategory.Value);
+                        db.CategoryItems.FirstOrDefault(c => c.CategoryItemID == selectedCategoryID).items.Add(itemHelper.item);
+                    }
+                    db.SaveChanges();
+                    return RedirectToAction("Details", new { id = itemHelper.item.ItemID });
+                }
+
+                return View(itemHelper.item);
+            }
+            else
+                return RedirectToAction("Index", "Home", new { message = "only worker can do this" });
         }
 
         //
@@ -301,12 +318,17 @@ namespace BeerShop.Controllers
 
         public ActionResult Delete(int id = 0)
         {
-            Item item = db.Items.Find(id);
-            if (item == null)
+            if (User is Worker || isWOrker == true)
             {
-                return HttpNotFound();
+                Item item = db.Items.Find(id);
+                if (item == null)
+                {
+                    return HttpNotFound();
+                }
+                return View(item);
             }
-            return View(item);
+            else
+                return RedirectToAction("Index", "Home", new { message = "only worker can do this" });
         }
 
         //
@@ -315,18 +337,23 @@ namespace BeerShop.Controllers
         [HttpPost, ActionName("Delete")]
         public ActionResult DeleteConfirmed(int id)
         {
-            Item item = db.Items.Find(id);
-            db.Entry(item).Collection(i => i.categories).Load();
-
-            item.categories.ToList().ForEach(cat => item.categories.Remove(cat));
-            var comments = db.Comments;
-            foreach (var c in item.comments)
+            if (User is Worker || isWOrker == true)
             {
-                comments.Remove(c);
+                Item item = db.Items.Find(id);
+                db.Entry(item).Collection(i => i.categories).Load();
+
+                item.categories.ToList().ForEach(cat => item.categories.Remove(cat));
+                var comments = db.Comments;
+                foreach (var c in item.comments)
+                {
+                    comments.Remove(c);
+                }
+                db.Items.Remove(item);
+                db.SaveChanges();
+                return RedirectToAction("Index");
             }
-            db.Items.Remove(item);
-            db.SaveChanges();
-            return RedirectToAction("Index");
+            else
+                return RedirectToAction("Index", "Home", new { message = "only worker can do this" });
         }
 
         protected override void Dispose(bool disposing)
